Implement heterogeneous lookup based, stream and connection containers:
- The stream/connection containers is now heterogeneous lookup based and
stores the streams as shared_ptr<void>. This is to be able to store the deleter
along with the key type. The hashing is done based on raw pointer value, and the
container supports lookups by both shared_ptr and void*, allowing seamless
operation in callback functions.
- Fixed invalid pointer passed to the StreamCallback log functions
- Added thunk functions for StreamCallback, ClientConnectionCallback,
ServerConnectionCallback, ServerListenerCallback that validates the parameters
and properly casts them to the correct types before passing them down.
- Added QUIC_API_ENABLE_PREVIEW_FEATURES macro definition to msquic dependency
- Added `-fdiagnostics-absolute-paths` to clang compiler flags
- Renamed msquic_api to msquic_context to prevent confusion
- Added shared_ptr_raw_{equal/hash}

Signed-off-by: Mustafa Kemal Gilor <hello@mkg.dev>