[project]: fix double-free for stream and conn
The callback handlers for stream/connection shutdown events weren't
checking the 'AppCloseInProgress' flag, which indicates that the close
is initated by the app itself (i.e., release of the shared_ptr) which
leads to the double-free of the connection/stream map elements. This is
now fixed.

Also, numerous fixes and tidy-ups were made:

- [msquic_base] grouped event decltypes into a struct
- [msquic_base] add a failure handler to StreamCallbackStartComplete
- [msquic_base] implemented close_stream function
- [flatbuffers/schemas] moved schemas to their own folder. headers are
now under mad/nexus/schemas path.
- [sample] moved sample applications to their own folder.
- [sample] sample applications now accept certificate path as cli arg
- [quic_configuration] added role parameter (client/server)
- [msquic_test_server] stop and join worker threads before exit

Signed-off-by: Mustafa Kemal Gilor <hello@mkg.dev>