[msquic] replace C++ API usage with C (#23)
I'm finally fed up with the C++ API because it's horrible. Reverted all
the changes, and the code now *again* properly uses the C API to do the
stuff needed to operate the MSQUIC.

- Removed the global MSQUIC API object, because it's a stupid idea.
- The MSQUIC API object is now collectively owned by the msquic_application
instances. msquic_application factory function automatically initializes
the API and hands it over to the first msquic_application object alive.
The subsequent objects will use the same object.
- Properly forward declared the MSQUIC types for msquic_* class smart
pointer types
- Removed quic_configuration member from the quic_application as it is
only needed during the initialization phase, which is now handled in the
factory method itself.
- msquic_application is now dependency injection friendly.
- Moved msquic_application const member from msquic_client/msquic_server
to msquic_base
- [msquic_client] listen function now takes alpn and port as argument.
- [msquic_client] implemented disconnect function.
- [vendor] removed msquic 2.3.6 and msquic.hpp from 2.4.4/include

Signed-off-by: Mustafa Kemal Gilor <hello@mkg.dev>